refactor: Server Actionsの分離とJSDocコメントの追加
